Added UI Settings Tests
